web: move InjectExtensions to web::utils::extensions
