+++ src/web/utils/mod.rs
@@ -1,0 +1,3 @@
+mod extensions;
+
+pub(crate) use self::extensions::InjectExtensions;
